[Java] Inline padding classes.
