+++ benchmarks-aeron/src/main/java/uk/co/real_logic/benchmarks/aeron/remote/EchoMessageTransceiver.java
@@ -20,6 +20,7 @@ import io.aeron.driver.MediaDriver;
 import org.agrona.concurrent.UnsafeBuffer;
 import uk.co.real_logic.benchmarks.remote.Configuration;
 import uk.co.real_logic.benchmarks.remote.MessageRecorder;
+import uk.co.real_logic.benchmarks.remote.MessageTransceiver;
 
 import java.util.Arrays;
 
@@ -30,6 +31,35 @@ import static org.agrona.BitUtil.SIZE_OF_LONG;
 import static org.agrona.BufferUtil.allocateDirectAligned;
 import static org.agrona.CloseHelper.closeAll;
 import static uk.co.real_logic.benchmarks.aeron.remote.AeronUtil.*;
+
+abstract class EchoMessageTransceiverProducerState extends MessageTransceiver
+{
+    UnsafeBuffer offerBuffer;
+    ExclusivePublication[] publications;
+    ExclusivePublication[] passivePublications;
+    long keepAliveIntervalNs;
+    long timeOfLastKeepAliveNs;
+    int sendIndex;
+
+    EchoMessageTransceiverProducerState(final MessageRecorder messageRecorder)
+    {
+        super(messageRecorder);
+    }
+}
+
+@SuppressWarnings("unused")
+abstract class EchoMessageTransceiverProducerStatePadded extends EchoMessageTransceiverProducerState
+{
+    byte p000, p001, p002, p003, p004, p005, p006, p007, p008, p009, p010, p011, p012, p013, p014, p015;
+    byte p016, p017, p018, p019, p020, p021, p022, p023, p024, p025, p026, p027, p028, p029, p030, p031;
+    byte p032, p033, p034, p035, p036, p037, p038, p039, p040, p041, p042, p043, p044, p045, p046, p047;
+    byte p048, p049, p050, p051, p052, p053, p054, p055, p056, p057, p058, p059, p060, p061, p062, p063;
+
+    EchoMessageTransceiverProducerStatePadded(final MessageRecorder messageRecorder)
+    {
+        super(messageRecorder);
+    }
+}
 
 public final class EchoMessageTransceiver extends EchoMessageTransceiverProducerStatePadded
 {
