--- benchmarks-aeron/src/main/java/uk/co/real_logic/benchmarks/aeron/remote/EchoMessageTransceiverProducerStatePadded.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2015-2020 Real Logic Limited.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package uk.co.real_logic.benchmarks.aeron.remote;
-
-import uk.co.real_logic.benchmarks.remote.MessageRecorder;
-
-@SuppressWarnings("unused")
-abstract class EchoMessageTransceiverProducerStatePadded extends EchoMessageTransceiverProducerState
-{
-    byte p000, p001, p002, p003, p004, p005, p006, p007, p008, p009, p010, p011, p012, p013, p014, p015;
-    byte p016, p017, p018, p019, p020, p021, p022, p023, p024, p025, p026, p027, p028, p029, p030, p031;
-    byte p032, p033, p034, p035, p036, p037, p038, p039, p040, p041, p042, p043, p044, p045, p046, p047;
-    byte p048, p049, p050, p051, p052, p053, p054, p055, p056, p057, p058, p059, p060, p061, p062, p063;
-
-    EchoMessageTransceiverProducerStatePadded(final MessageRecorder messageRecorder)
-    {
-        super(messageRecorder);
-    }
-}
